Change proxy handling to allow multiple proxies to be specified in package.json.
