+++ packages/react-scripts/scripts/start.js
@@ -31,7 +31,6 @@ const paths = require('../config/paths');
 const config = require('../config/webpack.config.dev');
 const devServerConfig = require('../config/webpackDevServer.config');
 const createWebpackCompiler = require('./utils/createWebpackCompiler');
-const addWebpackMiddleware = require('./utils/addWebpackMiddleware');
 
 const useYarn = fs.existsSync(paths.yarnLockFile);
 const cli = useYarn ? 'yarn' : 'npm';
@@ -71,9 +70,6 @@ function run(port) {
 
   // Serve webpack assets generated by the compiler over a web sever.
   const devServer = new WebpackDevServer(compiler, devServerConfig);
-
-  // Our custom middleware proxies requests to /index.html or a remote API.
-  addWebpackMiddleware(devServer);
 
   // Launch WebpackDevServer.
   devServer.listen(port, err => {
